requirements.txt
appoptics-apm==5.2.0
boto3==1.21.32
celery==5.2.5
colorlog==6.6.0
dacite==1.6.0
Django==4.0.3
django-csp==3.7
django-environ==0.8.1
git+https://github.com/pkarman/django-include-strip-tag@70ba69f#egg=django-include-strip-tag
git+https://github.com/foundertherapy/django-redis-secure@support-python-3.7#egg=django-redis-secure
gunicorn==20.1.0
hiredis==2.0.0
git+https://github.com/pkarman/jsonref.git@590c416#egg=jsonref
jsonschema[format]==4.4.0
jwcrypto==1.0
git+https://github.com/trussworks/kombu-fernet-serializers@lazy-fernet#egg=kombu-fernet-serializers
launchdarkly-server-sdk==7.4.0
git+https://github.com/trussworks/logindotgov-oidc-py.git@7cc5218#egg=logindotgov-oidc
mysqlclient==2.1.0
pyjwt==2.3.0
python-dateutil==2.8.2
redis==4.2.2
request-id-django-log==0.2.0
requests==2.27.1
rfc3339-validator==0.1.4
whitenoise==6.0.0
